Add seperate test site
